feat(singularity): implement HNSW and LSH ANN indices (ADR-0068)
- Added `AnnIndex` trait and `IndexBackend` enum in `src/index/mod.rs`.
- Implemented `BruteForce`, `HnswIndex` (via `hnsw_rs`), and `LshIndex` backends.
- Integrated ANN index into `Singularity` core with automatic synchronization on CRUD ops.
- Added database migration v7 to persist serialized index state in `csm_hnsw_graph`.
- Extended `FrameworkBuilder` to support opt-in selection of ANN backends.
- Added `HVec10240::set_bit` for easier vector manipulation.
- Fixed WASM build and stubbed persistence by adding missing `save_index`/`load_index` methods.
- Resolved type annotation ambiguity for `Arc` in search path.
- Simplified HNSW serialization to return empty vec and rely on rebuild fallback.
- Optimized `Singularity` and `Framework` for Clippy and restored retrieval statistics.
- Verified with integration tests for search accuracy and persistence roundtrip.
- Addressed scale requirements (>200k) with 1M capacity HNSW default.
