+++ llms-full.txt
@@ -433,6 +433,7 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub mod lsh
 - pub struct IndexStats
 - pub enum IndexBackend
+- impl Default for IndexBackend
 - pub trait AnnIndex
 
 ### src/lib.rs
@@ -462,13 +463,13 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub mod metadata_filter
 - pub mod semantic_triples
 - pub use metadata_filter::MetadataFilter
-- pub mod index
 - pub mod persistence
 - pub mod persistence_wasm
 - pub mod reservoir
 - pub mod retrieval
 - pub mod semantic_bridge
 - pub mod singularity
+- pub mod index
 - pub use crate::persistence_wasm as persistence
 - pub mod persistence
 - pub struct persistence::Persistence
@@ -537,10 +538,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 
 - pub struct Persistence
 - pub struct ConceptVersion
-- impl Persistence
-
-### src/persistence_index.rs
-
 - impl Persistence
 
 ### src/persistence_migrations.rs
@@ -3058,8 +3055,6 @@ Batch similarity computation tuned for cache locality and Rayon overhead.
 
 ## src/hyperdim_simd.rs
 
-## src/hyperdim_tests.rs
-
 ## src/index/brute_force.rs
 
 ### BruteForce
@@ -3131,7 +3126,7 @@ impl std::fmt::Debug for HnswIndex {
 ### LshIndex
 
 ```rust
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug)]
 pub struct LshIndex {
 }
 ```
@@ -3174,7 +3169,7 @@ Trait for Approximate Nearest Neighbor (ANN) indices.
 ### IndexBackend
 
 ```rust
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Default)]
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub enum IndexBackend {
     BruteForce,
     Hnsw { m: usize, ef_construction: usize, ef_search: usize },
@@ -3196,6 +3191,13 @@ pub struct IndexStats {
 ```
 
 Statistics for an ANN index.
+
+### impl Default for IndexBackend
+
+```rust
+impl Default for IndexBackend {
+}
+```
 
 ## src/lib.rs
 
@@ -3353,8 +3355,6 @@ impl Persistence {
     pub fn restore(&self, _path: &str) -> Result<()>;
     pub fn get_concept_history(&self, _id: &str, _limit: usize) -> Result<Vec<ConceptVersion>>;
     pub fn schema_version(&self) -> Result<i64>;
-    pub fn save_index(&self, _id: &str, _data: &[u8]) -> Result<()>;
-    pub fn load_index(&self, _id: &str) -> Result<Option<Vec<u8>>>;
     pub fn apply_migrations(&self, _target_version: i64) -> Result<()>;
 }
 ```
@@ -3693,15 +3693,6 @@ impl Persistence {
     pub fn delete_concept(&self, id: &str) -> Result<()>;
     pub fn save_association(&self, from: &str, to: &str, strength: f32) -> Result<()>;
     pub fn load_associations(&self, id: &str) -> Result<Vec<(String, f32)>>;
-}
-```
-
-## src/persistence_index.rs
-
-### impl Persistence
-
-```rust
-impl Persistence {
     pub fn checkpoint(&self) -> Result<()>;
     pub fn save_index(&self, id: &str, data: &[u8]) -> Result<()>;
     pub fn load_index(&self, id: &str) -> Result<Option<Vec<u8>>>;
@@ -3795,8 +3786,6 @@ impl Persistence {
     pub fn checkpoint(&self) -> Result<()>;
     pub fn health_check(&self) -> Result<()>;
     pub fn size(&self) -> Result<u64>;
-    pub fn save_index(&self, _id: &str, _data: &[u8]) -> Result<()>;
-    pub fn load_index(&self, _id: &str) -> Result<Option<Vec<u8>>>;
 }
 ```
 
@@ -4380,9 +4369,14 @@ impl Default for SingularityConfig {
 impl Singularity {
     pub fn new() -> Self;
     pub fn with_config(config: SingularityConfig) -> Self;
+    pub fn with_config_and_backend(config: SingularityConfig, backend: IndexBackend) -> Self;
     pub fn inject(&mut self, concept: Concept) -> Result<()>;
     pub fn get(&self, id: &str) -> Option<&Concept>;
     pub fn delete(&mut self, id: &str) -> Result<()>;
+    pub fn clear(&mut self);
+    pub fn update(&mut self, id: &str, new_vector: HVec10240) -> Result<()>;
+    pub fn find_similar(&self, query: &HVec10240, top_k: usize) -> Vec<(String, f32)>;
+    pub fn find_similar_arc(&self, query: &HVec10240, top_k: usize) -> Arc<[(String, f32)]>;
     pub fn find_similar_cached(&self, query: &HVec10240, top_k: usize) -> Arc<[(String, f32)]>;
     pub fn associate(&mut self, from: &str, to: &str, strength: f32) -> Result<()>;
     pub fn get_associations(&self, id: &str) -> Vec<(String, f32)>;
@@ -4429,11 +4423,6 @@ impl QueryCache {
 
 ```rust
 impl Singularity {
-    pub fn with_config_and_backend(config: SingularityConfig, backend: IndexBackend) -> Self;
-    pub fn clear(&mut self);
-    pub fn update(&mut self, id: &str, new_vector: HVec10240) -> Result<()>;
-    pub fn find_similar(&self, query: &HVec10240, top_k: usize) -> Vec<(String, f32)>;
-    pub fn find_similar_arc(&self, query: &HVec10240, top_k: usize) -> Arc<[(String, f32)]>;
     pub fn bundle_concepts_strict(&self, ids: &[String]) -> Result<HVec10240>;
     pub fn disassociate(&mut self, from: &str, to: &str) -> Result<()>;
     pub fn clear_associations(&mut self, id: &str) -> Result<()>;
