+++ src/lib.rs
@@ -77,8 +77,6 @@ pub mod index;
 #[cfg(all(not(target_arch = "wasm32"), feature = "persistence"))]
 pub mod persistence;
 #[cfg(all(not(target_arch = "wasm32"), feature = "persistence"))]
-mod persistence_index;
-#[cfg(all(not(target_arch = "wasm32"), feature = "persistence"))]
 mod persistence_migrations;
 #[cfg(all(not(target_arch = "wasm32"), feature = "persistence"))]
 mod persistence_ops;
